% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cmi_matrix}
\alias{cmi_matrix}
\title{Compute the matrix of CMI}
\usage{
cmi_matrix(data, labels, discrete = TRUE, disc = "equalwidth", nbins = 25,
  method = "emp")
}
\arguments{
\item{data}{matrix containing the OTU table}

\item{labels}{vector of discrete labels}

\item{discrete}{discretize the data. default is TRUE}

\item{disc}{method of performing discretization. available: equalfreq,
equalwidth, globalequalwidth. the default is equalwidth}

\item{nbins}{number of bins to discretize into}

\item{method}{method of performing mutual information calculation.}
}
\value{
returns a matrix with the pairwise conditional mutual information
}
\description{
Calculate the pairwise conditional mutual information for each of the
features in data.
}

