% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filter_otus}
\alias{filter_otus}
\title{Filter OTU maktrix by their abundance}
\usage{
filter_otus(biom_df, lvl = 0.2)
}
\arguments{
\item{biom_df}{data.frame that has be OTU data}

\item{lvl}{percentage of abundance to keep}
}
\value{
return the biom_df with a subset of OTUs
}
\description{
Many of the overall abundances for different taxa are quite low, so in this
function, we filter them to a smaller subset that contain only the variables
that are the most abundant up to some pre-determined level.
}

