% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_biom_matrix}
\alias{load_biom_matrix}
\title{Read in a file in BIOM format}
\usage{
load_biom_matrix(biom_fp)
}
\arguments{
\item{biom_fp}{string that points to the path of the biom file}
}
\value{
returns a data.frame that has the result from the biom & dense matrix
}
\description{
This function reads in a biom file using the BIOM utilities. The function
then converts the data from a sparse to dense format.
}
\author{
Gregory Ditzler
}

