% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load_metadata}
\alias{load_metadata}
\title{Load a map file in tab-delimited format}
\usage{
load_metadata(map_fp)
}
\arguments{
\item{map_fp}{string pointing to the path of the map file.}
}
\value{
data.frame containing the metadata
}
\description{
Load a map file in tab-delimited format
}
\author{
Gregory Ditzler
}

