% Generated by roxygen2 (4.0.2): do not edit by hand
\name{measure_otu_mi}
\alias{measure_otu_mi}
\title{Measure the mutual information in the OTU table}
\usage{
measure_otu_mi(biom_df, labels, discrete = FALSE, disc = "equalfreq",
  nbins = 25, method = "emp")
}
\arguments{
\item{biom_df}{data.frame containing the OTU table}

\item{labels}{class labels for each of the samples in the OTU table}

\item{discrete}{discretize the data. default is FALSE}

\item{disc}{method of performing discretization. available: equalfreq,
equalwidth, globalequalwidth. the default is equalfreq}

\item{nbins}{number of bins to discretize into}

\item{method}{method of performing mutual information calculation.}
}
\value{
returns a vector with the mutual information
}
\description{
This function measures the mutual information shared between the OTU
variables and the class labels that have been extracted from the map file
describing the metadata.
}

