% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mi_matrix}
\alias{mi_matrix}
\title{Compute the matrix of MI}
\usage{
mi_matrix(data, discrete = TRUE, disc = "equalwidth", nbins = 25,
  method = "emp")
}
\arguments{
\item{data}{matrix containing the OTU table}

\item{discrete}{discretize the data. default is TRUE}

\item{disc}{method of performing discretization. available: equalfreq,
equalwidth, globalequalwidth. the default is equalwidth}

\item{nbins}{number of bins to discretize into}

\item{method}{method of performing mutual information calculation.}
}
\value{
returns a matrix with the pairwise mutual information
}
\description{
Calculate the pairwise mutual information for each of the features in
data.
}

