% Generated by roxygen2 (4.0.2): do not edit by hand
\name{retrieve_labels}
\alias{retrieve_labels}
\title{Extract labels from a map data.frame}
\usage{
retrieve_labels(map_df, sample_ids, col_name)
}
\arguments{
\item{map_df}{map data.frame from load_metadata}

\item{sample_ids}{list of sample ids containing the order of the samples
in the data matrix}

\item{col_name}{column name in the map file to extract the labels}
}
\description{
Extract labels from a map data.frame
}
\author{
Gregory Ditzler
}

