% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scale_matrix}
\alias{scale_matrix}
\title{Scale a communuity data matrix}
\usage{
scale_matrix(biom_df, rescale = TRUE)
}
\arguments{
\item{biom_df}{data.frame from loading the biom file.}

\item{recscale}{if FALSE observations sum to one; otherwise they are rescaled
to produce integer values. default is TRUE}
}
\value{
return the scaled matrix
}
\description{
This function normalizes the community data matrix, or sometimes referred to
as the OTU table. This needs to be performed because multiple samples could
have the same proportion of bacteria; however, the abundances are different.
For example, a=[1,2,3] is proportionally equal to b=[3,6,9]. The matrix is
scaled and written to df$data_dense.
}
\author{
Gregory Ditzler
}

