% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter_otus.R
\name{filter_otus}
\alias{filter_otus}
\title{Filter OTU matrix by their abundance}
\usage{
filter_otus(biom_df, lvl = 0.2)
}
\arguments{
\item{biom_df}{data.frame that has be OTU data}

\item{lvl}{percentage of abundance to keep}
}
\value{
return the lists containing the reduced data, sample ids, feature names
and feature ids.
}
\description{
Many of the overall abundances for different taxa are quite low, so in this
function, we filter them to a smaller subset that contain only the variables
that are the most abundant up to some pre-determined level.
}
\author{
Gregory Ditzler
}

