% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mi_selection}
\alias{mi_selection}
\title{Mutual Information Maximization}
\usage{
mi_selection(biom_df, labels, n_select, discrete = FALSE,
  disc = "equalfreq", nbins = 25, method = "emp")
}
\arguments{
\item{biom_df}{data.frame containing the OTU table}

\item{labels}{class labels for each of the samples in the OTU table}

\item{n_select}{number of OTUs to select}

\item{discrete}{discretize the data. default is FALSE}

\item{disc}{method of performing discretization. available: equalfreq,
equalwidth, globalequalwidth. the default is equalfreq}

\item{nbins}{number of bins to discretize into}

\item{method}{method of performing mutual information calculation.}
}
\value{
returns a data.frame containing the highest ranking indices
according to mutual information
}
\description{
Mutual Information Maximization
}
\author{
Gregory Ditzler
}

